user-service/requirements.txt
Flask==2.3.3
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0